[Refactor]: constant 폴더 정리 및 pb collection 상수화
