+++ frontend/server-actions/categories/category.ts
@@ -4,6 +4,7 @@ import client from '@/libs/pocketbase';
 import { revalidateTag } from 'next/cache';
 import { REVALIDATE_TAG } from '@/constants/constants';
 import { ICategory } from '@/types/category/type';
+import { PB_COLLECTIONS } from '@/constants/constants';
 
 interface ICreateAction {
   data: { name: string };
@@ -20,13 +21,13 @@ interface IUpdateAction {
 
 const createCategory = async ({ data }: ICreateAction) => {
   await client
-    .collection('categories')
+    .collection(PB_COLLECTIONS.CATEGORIES)
     .create(data)
     .then(() => revalidateTag(REVALIDATE_TAG.CATEGORY));
 };
 
 const getCategories = async () => {
-  const categories: ICategory[] = await client.collection('categories').getFullList({
+  const categories: ICategory[] = await client.collection(PB_COLLECTIONS.CATEGORIES).getFullList({
     sort: 'created',
     cache: 'no-store',
     next: { tags: [REVALIDATE_TAG.CATEGORY], revalidate: 10 },
@@ -44,7 +45,7 @@ const getCategories = async () => {
 // TODO - 공통으로 사용 가능할 것으로 보임
 const deleteCategory = async ({ id }: IDeleteAction) => {
   await client
-    .collection('categories')
+    .collection(PB_COLLECTIONS.CATEGORIES)
     .delete(id)
     .then(() => revalidateTag(REVALIDATE_TAG.CATEGORY));
 };
@@ -52,7 +53,7 @@ const deleteCategory = async ({ id }: IDeleteAction) => {
 // TODO - 공통으로 사용 가능할 것으로 보임
 const updateCategory = async ({ id, name }: IUpdateAction) => {
   await client
-    .collection('categories')
+    .collection(PB_COLLECTIONS.CATEGORIES)
     .update(id, { name })
     .then(() => revalidateTag(REVALIDATE_TAG.CATEGORY));
 };
